Delete translation utils
These were used by the wagtail-localize-pontoon module. But they will be
reimplemented as methods of the new Translation model when I add that in
a future PR.
