+++ wagtail_localize/translation/tests/test_translatablerevision.py
@@ -15,10 +15,25 @@ from wagtail_localize.translation.models import (
     MissingTranslationError,
     MissingRelatedObjectError,
     TranslationContext,
+    SegmentLocation,
+    TemplateLocation,
+    RelatedObjectLocation,
 )
-from wagtail_localize.translation.utils import insert_segments
-from wagtail_localize.translation.segments import RelatedObjectValue
+from wagtail_localize.translation.segments import TemplateValue, RelatedObjectValue
 from wagtail_localize.translation.segments.extract import extract_segments
+
+
+def insert_segments(revision, locale, segments):
+    """
+    Inserts the list of untranslated segments into translation memory
+    """
+    for segment in segments:
+        if isinstance(segment, TemplateValue):
+            TemplateLocation.from_template_value(revision, segment)
+        elif isinstance(segment, RelatedObjectValue):
+            RelatedObjectLocation.from_related_object_value(revision, segment)
+        else:
+            SegmentLocation.from_segment_value(revision, locale, segment)
 
 
 def create_test_page(**kwargs):
